Refactor the markdown node renderers
