+++ lib/Renderer.js
@@ -1,30 +1,12 @@
 var marked = require('marked');
 var _ = require('lodash');
 var RendererUtils = require('./RendererUtils');
-var multiline = require('multiline-slash');
-var Prism = require('prismjs');
+var renderHeading = require('./Renderer__renderHeading');
+var CodeRenderer = require('./Renderer__renderCode');
+var LinkRenderer = require('./Renderer__renderLink');
 var assign = _.assign;
 
-var AnchorableHeadingTmpl = _.template(
-  multiline(function() {;
-    // <h<%- level %> class="anchorable-heading">
-    //   <a name="<%- id %>" class="anchorable-heading__anchor"></a>
-    //   <a href="#<%- id %>" class="anchorable-heading__link icon icon-link"></a><span class="anchorable-heading__text"><%= text %></span>
-    // </h<%- level %>>
-  })
-);
-
-var HeadingTmpl = _.template(
-  multiline(function() {;
-    // <h<%- level %>>
-    //   <%= text %>
-    // </h<%- level %>>
-  })
-);
-
-var RE_INTERNAL_LINK = Object.freeze(/^mega:\/\//);
-
-function createRenderer(config) {
+function Renderer(config) {
   var renderer = new marked.Renderer();
   var markedOptions = Object.freeze({
     renderer: renderer,
@@ -36,145 +18,89 @@ function createRenderer(config) {
     pedantic: false,
   });
 
-  var syntaxHighlighting = assign({
-    languages: [],
-    aliases: {}
-  }, config.syntaxHighlighting);
-
-  var state;
-  var runOptions;
+  var runState, runOptions;
   var inSinglePageMode = config.layoutOptions.singlePageMode;
   var codeBlockRenderers = {};
+  var renderCode = CodeRenderer(config.syntaxHighlighting);
+  var renderLink = LinkRenderer(config);
 
-  syntaxHighlighting.languages.forEach(function(name) {
-    try {
-      require('prismjs/components/prism-' + name);
-    }
-    catch(e) {
-      console.warn("SyntaxHighlighter: definition for language '%s' could not be found",
-        name
-      );
-    }
-  });
-
-  function createState(baseURL) {
-    return { baseURL: baseURL || '', toc: [] };
+  function createRunState(baseURL) {
+    return { baseURL: inSinglePageMode ? (baseURL || '') : null, toc: [] };
   }
 
   function createRunOptions(options) {
-    options = options || {};
-
-    return { anchorableHeadings: options.anchorableHeadings !== false };
+    return { anchorableHeadings: !options || options.anchorableHeadings !== false };
   }
 
   // this could be heavily optimized, but meh for now
   renderer.heading = function(text, level) {
-    var scopedId = RendererUtils.normalizeHeading(
-      RendererUtils.htmlToText(text.split('\n')[0])
-    );
-
-    var id = inSinglePageMode ? joinBySlash(state.baseURL, scopedId) : scopedId;
-
-    state.toc.push({
-      id: id,
-      scopedId: scopedId,
-      level: level,
-      html: text,
-      text: RendererUtils.markdownToText(text).trim()
-    });
-
-    if (runOptions.anchorableHeadings && id && id.length) {
-      return AnchorableHeadingTmpl({
-        // we need to strip any leading # because in SinglePageMode all baseURL
-        // values will have that
-        id: stripLeadingHash(id),
-        level: level,
-        text: text,
-      });
-    }
-    else {
-      return HeadingTmpl({ level: level, text: text });
-    }
+    return renderHeading(text, level, runState, runOptions);
   };
 
-  renderer.link = function(srcHref, title, text) {
-    var href = srcHref.replace(RE_INTERNAL_LINK, '');
-    var tagString = '<a href="' + href + '"';
-
-    if (title) {
-      tagString += ' title="' + _.escape(title) + '"';
-    }
-
-    if (href === srcHref && href[0] !== '#' && config.launchExternalLinksInNewTabs) {
-      tagString += ' target="_blank"';
-    }
-
-    return tagString + '>' + text + '</a>';
-  };
+  renderer.link = renderLink;
 
   renderer.code = function(code, language) {
     if (codeBlockRenderers[language]) {
       return codeBlockRenderers[language](code);
     }
     else {
-      var languageCode = syntaxHighlighting.aliases[language] || language;
-      var grammar = Prism.languages[languageCode];
-      var html = code;
-
-      if (grammar) {
-        html = Prism.highlight(code, grammar);
-      }
-
-      return (
-        '<pre class="language-' + (grammar ? languageCode : 'none') + '"><code>' +
-          html +
-        '</code></pre>'
-      );
+      return renderCode(code, language);
     }
   };
 
+  /**
+   * Render Markdown to HTML.
+   *
+   * @param {String} text
+   *        Markdown text.
+   *
+   * @param {Object} options
+   * @param {String} [options.baseURL=null]
+   *        The url to prefix the anchor links with. This is necessary only in
+   *        the Single Page Mode because all anchors must be absolute then. In
+   *        the regular mode, the anchors are shortened to be more friendly
+   *        since they assume there won't be conflicts (because each "page" is
+   *        expected to represent a single document.)
+   *
+   *        Pass the documentNode's @meta.href value here if you're compiling
+   *        from a corpus node.
+   *
+   * @param {Boolean} [options.anchorableHeadings=true]
+   *        Turn this off if you do not want the headings to have anchors -
+   *        the [name] attribute and the .anchorable-heading stuff.
+   *
+   * @param {Boolean} [options.withTOC=false]
+   *        Turn this on if you want the ToC meta-data. The return value will
+   *        be an object of the shape: `{ html: String, toc: Array.<Object> }`.
+   *
+   * @return {String|Object}
+   *         The HTML.
+   */
   var exports = function renderMarkdown(text, options) {
-    var html;
-
-    options = options || {};
-    state = createState(options.baseURL);
-    runOptions = createRunOptions(options);
-
-    html = marked(text, assign({}, markedOptions, {
-      sanitize: options.sanitize !== false
-    }));
-
-    if (options && options.trimHTML) {
-      html = RendererUtils.trimHTML(html);
-    }
-
-    state.baseURL = '';
-    runOptions = createRunOptions();
-
-    return html;
-  };
-
-  exports.withTOC = function(text, options) {
     var html, toc;
 
     options = options || {};
-    state = createState(options.baseURL);
+    runState = createRunState(options.baseURL);
     runOptions = createRunOptions(options);
 
     html = marked(text, assign({}, markedOptions, {
       sanitize: options.sanitize !== false
     }));
 
-    toc = state.toc;
+    toc = runState.toc;
 
     if (options && options.trimHTML) {
       html = RendererUtils.trimHTML(html);
     }
 
-    state = createState();
+    runState = createRunState();
     runOptions = createRunOptions();
 
-    return { html: html, toc: toc };
+    return options.withTOC ? { html: html, toc: toc } : html;
+  };
+
+  exports.withTOC = function(text, options) {
+    return exports(text, assign({ withTOC: true }, options));
   };
 
   /**
@@ -200,14 +126,4 @@ function createRenderer(config) {
   return exports;
 }
 
-module.exports = createRenderer;
-
-function stripLeadingHash(x) {
-  return x[0] === '#' ? x.slice(1) : x;
-}
-
-function joinBySlash(a, b) {
-  var shouldDelimit = a && a[a.length-1] !== '/';
-
-  return a + (shouldDelimit ? '/' : '') + b;
-}
+module.exports = Renderer;
